requirements.txt
python-twitter==3.5
slackclient==2.8.1
